analysis_automation/requirements.txt
asgiref==3.6.0
click==8.1.3
Flask==2.2.3
fpdf==1.7.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paho-mqtt==1.6.1
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
